requirements.txt
beautifulsoup4==4.8.2
bottle==0.12.19
numpy==1.18.1
requests==2.22.0
tabula-py==2.0.4
tornado==6.0.3
